add --prefork arg and change version v1.5.0

+++ main.go
@@ -39,6 +39,7 @@ var embedDirStatic embed.FS
 func main() {
 
 	arg_port := flag.Int("port", 8000, "Change default listen port")
+	arg_prefork := flag.Bool("prefork", false, "Enable spawn multiple processes")
 
 	arg_user := flag.String("user", "", "Login for user basic auth")
 	arg_password := flag.String("password", "", "Password for user basic auth")
@@ -60,12 +61,13 @@ func main() {
 
 		inf := []string{
 			``,
-			`v1.4.0`,
+			`v1.5.0`,
 			``,
 			`usage: http-here [options] [path]`,
 			``,
 			`options:`,
 			`     --port                    Port to use [8000]`,
+			`     --prefork                 Enable spawn multiple processes`,
 			``,
 			`     --user                    Login for basic authorization.`,
 			`     --password                Password for basic authorization.`,
@@ -130,7 +132,7 @@ func main() {
 	engine := html.NewFileSystem(http.FS(view_fs), ".html")
 
 	config := fiber.Config{
-		Prefork:               false,
+		Prefork:               *arg_prefork,
 		DisableStartupMessage: true,
 		ServerHeader:          "",
 		Views:                 engine,
@@ -159,22 +161,6 @@ func main() {
 		return c.Next()
 	})
 
-	/*
-	   app.Use([]string{"/api", "/"}, func(c *fiber.Ctx) error {
-
-	       c.Locals("arg_fold", arg_fold)
-
-
-	       if *arg_upload_disable {
-	   		c.Locals("arg_upload_disable", "1")
-	   	}
-	   	if *arg_folder_make_disable {
-	   		c.Locals("arg_folder_make_disable", "1")
-	   	}
-
-	       return c.Next()
-	   })*/
-
 	app.Use(cors.New(cors.Config{
 		//AllowOrigins: "*",
 		AllowCredentials: true,
@@ -189,26 +175,33 @@ func main() {
 	cian := color.New(color.FgCyan).SprintFunc()
 	yellow := color.New(color.FgYellow).SprintFunc()
 
-	if *arg_basic {
+	if *arg_basic && *arg_prefork {
+
+		fmt.Println("")
+		fmt.Println("You can not run --prefork with --basic. Use --user and --password instead.")
+		return
+	}
+
+	if *arg_basic && !*arg_prefork {
 
 		fmt.Println("")
 		fmt.Println("  Basic auth set: ")
 
-		var data = map[string]string{}
+		var password_list = map[string]string{}
 
 		for i := range 10 {
 
 			login := "login" + strconv.Itoa(i) + controller.RandStringRunes(2)
 			password := controller.RandStringRunes(16)
 
-			data[login] = password
+			password_list[login] = password
 
 			fmt.Println("         " + login + "    " + password)
 		}
 
 		app.Use(basicauth.New(basicauth.Config{
 
-			Users: data,
+			Users: password_list,
 
 			Unauthorized: func(c *fiber.Ctx) error {
 
@@ -241,8 +234,10 @@ func main() {
 			},
 		}))
 
-		fmt.Println("")
-		fmt.Println("  Basic auth set: " + cian(*arg_user) + " " + cian(*arg_password))
+		if !fiber.IsChild() {
+			fmt.Println("")
+			fmt.Println("  Basic auth set: " + cian(*arg_user) + " " + cian(*arg_password))
+		}
 	}
 
 	app.Use("/__assets", filesystem.New(filesystem.Config{
@@ -251,27 +246,29 @@ func main() {
 		Browse:     false,
 	}))
 
-	if _, err3 := os.Stat(filepath.Join(homepath, ".httphere", "temp")); err3 != nil {
+	if !fiber.IsChild() {
 
-		fmt.Println("")
-		fmt.Println("  Make temp folder")
+		if _, err3 := os.Stat(filepath.Join(homepath, ".httphere", "temp")); err3 != nil {
 
-		if err4 := os.MkdirAll(filepath.Join(homepath, ".httphere", "temp"), os.ModePerm); err4 != nil {
-			log.Fatal(err4)
+			fmt.Println("")
+			fmt.Println("  Make temp folder")
+
+			if err4 := os.MkdirAll(filepath.Join(homepath, ".httphere", "temp"), os.ModePerm); err4 != nil {
+				log.Fatal(err4)
+			}
+		} else {
+
+			fmt.Println("")
+			fmt.Println(yellow("  Clear temp folder"))
+
+			if err := os.RemoveAll(filepath.Join(homepath, ".httphere", "temp")); err != nil {
+				log.Fatal(err)
+			}
+
+			if err4 := os.MkdirAll(filepath.Join(homepath, ".httphere", "temp"), os.ModePerm); err4 != nil {
+				log.Fatal(err4)
+			}
 		}
-	} else {
-
-		fmt.Println("")
-		fmt.Println(yellow("  Clear temp folder"))
-
-		if err := os.RemoveAll(filepath.Join(homepath, ".httphere", "temp")); err != nil {
-			log.Fatal(err)
-		}
-
-		if err4 := os.MkdirAll(filepath.Join(homepath, ".httphere", "temp"), os.ModePerm); err4 != nil {
-			log.Fatal(err4)
-		}
-
 	}
 
 	//app.Static("/__temp", filepath.Join(homepath, ".httphere", "temp"))
@@ -352,7 +349,7 @@ func main() {
 		crt_is_exists = true
 	}
 
-	if *arg_tls && crt_is_exists {
+	if *arg_tls && crt_is_exists && !fiber.IsChild() {
 
 		fmt.Println("")
 		fmt.Println("  Crt: " + yellow(crt_filename))
@@ -360,7 +357,7 @@ func main() {
 		fmt.Println("")
 	}
 
-	if *arg_tls && !crt_is_exists {
+	if *arg_tls && !crt_is_exists && !fiber.IsChild() {
 
 		//log.Fatal(crt_is_exists)
 
@@ -429,50 +426,51 @@ set_var EASYRSA_CERT_EXPIRE 3650
 		fmt.Println("")
 	}
 
-	fmt.Println("")
-	if *arg_tls {
-		fmt.Println(yellow("  TLS Server port " + strconv.Itoa(*arg_port)))
-	} else {
-		fmt.Println("  Server port " + cian(strconv.Itoa(*arg_port)))
-	}
+	if !fiber.IsChild() {
+		fmt.Println("")
+		if *arg_tls {
+			fmt.Println(yellow("  TLS Server port " + strconv.Itoa(*arg_port)))
+		} else {
+			fmt.Println("  Server port " + cian(strconv.Itoa(*arg_port)))
+		}
 
-	fmt.Println("")
-	ifaces, err := net.Interfaces()
-	if err != nil {
-		fmt.Print(fmt.Errorf("localAddresses: %+v\n", err.Error()))
-		return
-	}
-	for _, i := range ifaces {
-		addrs, err := i.Addrs()
+		fmt.Println("")
+		ifaces, err := net.Interfaces()
 		if err != nil {
 			fmt.Print(fmt.Errorf("localAddresses: %+v\n", err.Error()))
-			continue
+			return
 		}
-		for _, a := range addrs {
-			switch v := a.(type) {
+		for _, i := range ifaces {
+			addrs, err := i.Addrs()
+			if err != nil {
+				fmt.Print(fmt.Errorf("localAddresses: %+v\n", err.Error()))
+				continue
+			}
+			for _, a := range addrs {
+				switch v := a.(type) {
 
-			case *net.IPNet:
-				//fmt.Printf("%v : %s [%v/%v]\n", i.Name, v, v.IP, v.Mask)
-				//fmt.Printf("%v \n", v.IP)
-				if v.IP.To4() != nil {
-					//fmt.Println( "yes, ipv4" )
+				case *net.IPNet:
 
-					if *arg_tls {
-						fmt.Println("     https://" + v.IP.String() + ":" + cian(strconv.Itoa(*arg_port)))
-					} else {
-						fmt.Println("     http://" + v.IP.String() + ":" + cian(strconv.Itoa(*arg_port)))
+					if v.IP.To4() != nil {
+
+						if *arg_tls {
+							fmt.Println("     https://" + v.IP.String() + ":" + cian(strconv.Itoa(*arg_port)))
+						} else {
+							fmt.Println("     http://" + v.IP.String() + ":" + cian(strconv.Itoa(*arg_port)))
+						}
 					}
 				}
+
 			}
-
 		}
-	}
-	fmt.Println("")
 
-	fmt.Println("  Serve folder: " + cian(arg_fold))
-	fmt.Println("")
-	fmt.Println(cian("  [ Control + C ] ") + "Break Server")
-	fmt.Println("")
+		fmt.Println("")
+
+		fmt.Println("  Serve folder: " + cian(arg_fold))
+		fmt.Println("")
+		fmt.Println(cian("  [ Control + C ] ") + "Break Server")
+		fmt.Println("")
+	}
 
 	if *arg_tls {
